Add ong controller
