+++ aulas/backend/src/routes.js
@@ -1,32 +1,9 @@
 const express = require('express')
-const crypto = require('crypto')
-const connection = require('./database/index')
-
 const routes = express.Router()
+const OngController = require('./controllers/ong-controller')
 
-routes.post('/ongs', async (req, res) => {
-  const { name, email, whatsapp, city, uf } = req.body
-  const id = crypto.randomBytes(4).toString('HEX')
-  const ong = {
-    id: id,
-    name: name,
-    email: email,
-    whatsapp: whatsapp,
-    city: city,
-    uf: uf
-  }
+routes.get('/ongs', OngController.list)
 
-  console.log(ong)
-
-  await connection('ongs').insert(ong)
-
-  return res.json({ id })
-})
-
-routes.get('/ongs', async (req, res) => {
-  const data = await connection.select().table('ongs')
-
-  return res.json(data)
-})
+routes.post('/ongs', OngController.create)
 
 module.exports = routes
